+++ packages/analytics/analytics-utilities/src/utils/color-overrides.ts
@@ -87,7 +87,7 @@ export const statusCodeGroupColors: ColorMap = {
 export const stateColors: Record<ColorState, string> = {
   error: tokens.KUI_COLOR_TEXT_DANGER,
   warning: tokens.KUI_COLOR_TEXT_WARNING_WEAK,
-  empty: tokens.KUI_COLOR_BACKGROUND_NEUTRAL_WEAK,
+  empty: '#9DA99D', // chart.series.9
   neutral: tokens.KUI_COLOR_BACKGROUND_NEUTRAL,
   '____OTHER____': tokens.KUI_COLOR_BACKGROUND_INFO,
 }