+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.6.56](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-metric-provider%4011.6.55...%40kong-ui-public%2Fanalytics-metric-provider%4011.6.56) (2026-09-24)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.6.55](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-metric-provider%4011.6.54...%40kong-ui-public%2Fanalytics-metric-provider%4011.6.55) (2026-09-24)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.6.54](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-metric-provider%4011.6.53...%40kong-ui-public%2Fanalytics-metric-provider%4011.6.54) (2026-09-24)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.6.53](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-metric-provider%4011.6.52...%40kong-ui-public%2Fanalytics-metric-provider%4011.6.53) (2026-09-24)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.6.52](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-metric-provider%4011.6.51...%40kong-ui-public%2Fanalytics-metric-provider%4011.6.52) (2026-09-24)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [11.6.51](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-metric-provider%4011.6.50...%40kong-ui-public%2Fanalytics-metric-provider%4011.6.51) (2026-09-21)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider